# openapi.yaml
openapi: 3.0.3
info:
  title: RemitWise Frontend API (v1)
  version: 1.0.0
  description: |
    This OpenAPI document describes the current v1 API surface for RemitWise.

    Versioning: This repo uses URL-prefix versioning (e.g. /api/v1/...). The existing
    API is considered v1. To avoid breaking existing clients, the Next.js app
    rewrites `/api/*` to `/api/v1/*`.

    Deprecation policy: When a new major version is released (for example v2),
    older major versions will be supported for a minimum of 6 months before
    removal. Security patches and critical bug fixes will be backported during
    the deprecation window. Migration guides should be provided for clients.
servers:
  - url: /api/v1
    description: v1 base path (local/dev)
  - url: /api
    description: legacy base path (rewrites to v1)
paths:
...
  /remittance/quote:
    get:
      summary: Get Remittance Quote
      tags:
        - Remittance
      parameters:
        - in: query
          name: amount
          schema:
            type: number
          required: true
          description: Amount to send (> 0)
          example: 100.5
        - in: query
          name: currency
          schema:
            type: string
          required: true
          description: 3-letter ISO code
          example: USD
        - in: query
          name: toCurrency
          schema:
            type: string
          required: true
          description: 3-letter ISO code
          example: PHP
      responses:
        '200':
          description: Quote successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemittanceQuoteResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/remittance/history:
    get:
      summary: Get Remittance History
      tags:
        - Remittance
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            maximum: 200
          example: 20
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum:
              - completed
              - failed
              - pending
          example: completed
      responses:
        '200':
          description: History records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemittanceHistoryResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /split/initialize:
    post:
      summary: Initialize Split Configuration
      tags:
        - Split
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplitPercentages'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SplitUpdateResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/insurance/{id}/pay:
    post:
      summary: Pay Insurance Premium
      tags:
        - Insurance
      security:
        - xUserHeader: []
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: ins_7890
      responses:
        '200':
          description: Returns XDR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/XdrResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: remitwise-session
      description: Iron Session cookie for authenticated users
    xUserHeader:
      type: apiKey
      in: header
      name: x-user
      description: Stellar public key of the caller
...
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: |
            A translated error message based on the Accept-Language header.
            If the translation is missing or language is not supported, defaults to English.
            Example translations:
             - en: "Unauthorized: missing or invalid x-user header"
             - es: "No autorizado: falta o es inválido el encabezado x-user"
          example: 'Unauthorized: missing or invalid x-user header'
...
    RemittanceQuoteResponse:
      type: object
      properties:
        sendAmount:
          type: number
          example: 100
        receiveAmount:
          type: number
          example: 5600.5
        fee:
          type: number
          example: 1.5
        rate:
          type: number
          example: 56
        expiry:
          type: string
          format: date-time
          example: '2023-10-12T07:20:50.52Z'
        source:
          type: string
          enum:
            - anchor
            - stellar
          example: anchor
...
    RemittanceHistoryResponse:
      type: object
      properties:
        userAddress:
          type: string
          example: GBBD47IF6LWK7P7MDEVSCWR7DPUWV3NY3DTQEVFL4NAT4AQH3ZLLFLA5
      additionalProperties: true
...
    SplitPercentages:
      type: object
      required:
        - spending
        - savings
        - bills
        - insurance
      properties:
        spending:
          type: number
          example: 50
        savings:
          type: number
          example: 25
        bills:
          type: number
          example: 20
        insurance:
          type: number
          example: 5
    SplitUpdateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        xdr:
          type: string
          example: AAAAAgAAAAA...
        simulate:
          type: object
          additionalProperties: true
        message:
          type: string
          example: Successfully updated split percentages
...
    XdrResponse:
      type: object
      properties:
        xdr:
          type: string
          example: AAAAAgAAAAA...
...